+++ brewcontrol/jvm/src/main/scala/brewcontrol/PidController.scala
@@ -13,11 +13,11 @@ import scala.concurrent.duration.FiniteDuration
  *
  * @see http://en.wikipedia.org/wiki/PID_controller#PID_controller_theory
  */
-class PidController(sp: Rx[Float], pv: Rx[Float], updateInterval: FiniteDuration)(implicit scheduler: Scheduler, ec: ExecutionContext) extends LazyLogging {
+class PidController(sp: Rx[Double], pv: Rx[Double], updateInterval: FiniteDuration)(implicit scheduler: Scheduler, ec: ExecutionContext) extends LazyLogging {
 
-  val Kp = 1f
+  val Kp = 1d
 
-  def calcOutput(): Float = {
+  def calcOutput(): Double = {
     val error = sp() - pv()
     val o = Kp * error
     logger.debug(s"PID calculation: o=$o, sp=${sp()}, pv=${pv()}, Kp=$Kp")
