+++ brewcontrol/shared/src/main/scala/brewcontrol/DTOs.scala
@@ -1,5 +1,5 @@
 package brewcontrol
 
-case class Reading(timestamp: Long, sensorId: String, name: String, value: Float)
+case class Reading(timestamp: Long, sensorId: String, name: String, value: Double)
 
 case class RelayState(name: String, value: Boolean)
